Add Abstract and Concrete ResultParameterProcessor Classes
The AbstractResultParameterProcessor class has the abstract
implementation of the methods required to convert the SQL types in
the result of database operations into ballerina types and other
methods that are related to that task.
The concrete ResultParameterProcessor class contains the
implementation of those methods with protected access modifier to
allow connector modules to extend the functionality if needed.